checks and validations
