+++ src/utils/botUtils.js
@@ -73,19 +73,16 @@ async function safeDM(user, message) {
 }
 
 const permissions = {
-  ADMINISTRATOR: "Administrator",
-  VIEW_AUDIT_LOG: "View audit log",
-  MANAGE_GUILD: "Manage server",
-  MANAGE_ROLES: "Manage roles",
-  MANAGE_CHANNELS: "Manage channels",
+  CREATE_INSTANT_INVITE: "Create instant invite",
   KICK_MEMBERS: "Kick members",
   BAN_MEMBERS: "Ban members",
-  CREATE_INSTANT_INVITE: "Create instant invite",
-  CHANGE_NICKNAME: "Change nickname",
-  MANAGE_NICKNAMES: "Manage nicknames",
-  MANAGE_EMOJIS: "Manage emojis",
-  MANAGE_WEBHOOKS: "Manage webhooks",
-  VIEW_CHANNEL: "View channels",
+  ADMINISTRATOR: "Administrator",
+  MANAGE_CHANNELS: "Manage channels",
+  MANAGE_GUILD: "Manage server",
+  VIEW_AUDIT_LOG: "View audit log",
+  PRIORITY_SPEAKER: "Priority speaker",
+  STREAM: "Video",
+  VIEW_CHANNEL: "View channel",
   SEND_MESSAGES: "Send messages",
   SEND_TTS_MESSAGES: "Send TTS messages",
   MANAGE_MESSAGES: "Manage messages",
@@ -94,16 +91,24 @@ const permissions = {
   READ_MESSAGE_HISTORY: "Read message history",
   MENTION_EVERYONE: "Mention everyone",
   USE_EXTERNAL_EMOJIS: "Use external emojis",
-  ADD_REACTIONS: "Add reactions",
+  VIEW_GUILD_INSIGHTS: "View server insights",
   CONNECT: "Connect",
   SPEAK: "Speak",
   MUTE_MEMBERS: "Mute members",
   DEAFEN_MEMBERS: "Deafen members",
   MOVE_MEMBERS: "Move members",
   USE_VAD: "Use voice activity",
-  PRIORITY_SPEAKER: "Priority speaker",
-  VIEW_GUILD_INSIGHTS: "View server insights",
-  STREAM: "Video",
+  CHANGE_NICKNAME: "Change nickname",
+  MANAGE_NICKNAMES: "Manage nicknames",
+  MANAGE_ROLES: "Manage roles",
+  MANAGE_WEBHOOKS: "Manage webhooks",
+  MANAGE_EMOJIS_AND_STICKERS: "Manage emojis and stickers",
+  USE_APPLICATION_COMMANDS: "Use Application Commands",
+  REQUEST_TO_SPEAK: "Request to Speak",
+  MANAGE_THREADS: "Manage Threads",
+  USE_PUBLIC_THREADS: "Use Public Threads",
+  USE_PRIVATE_THREADS: "Use Private Threads",
+  USE_EXTERNAL_STICKERS: "Use External Stickers",
 };
 
 module.exports = {
